requirements----.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
Faker
requests
pytest
pytest-cov
Werkzeug==0.16.0
flask-cors